golang-ssh: tweak session creation
Currently in order to create session we're going to dial remote twice -
one to check if it's available in a retry loop, second time to actually
create session.

This PR makes the client to dial only once. Also disables dial retries,
making it configurable.

Last but not least this PR drops dependency on github.com/docker/machine
package.